+++ src/is/tabbable.js
@@ -4,6 +4,7 @@
 import contextToElement from '../util/context-to-element';
 import tabindexValue from '../util/tabindex-value';
 import focusRelevant from './focus-relevant';
+import getFrameElement from '../util/get-frame-element';
 import platform from '../util/platform';
 import {getImageOfArea} from '../util/image-map';
 import {
@@ -40,6 +41,26 @@ function isTabbableRules({
     // The on-screen keyboard does not provide a way to focus the next input element (like iOS does).
     // That leaves us with no option to advance focus by keyboard, ergo nothing is tabbable (keyboard focusable).
     return false;
+  }
+
+  const frameElement = getFrameElement(element);
+  if (frameElement) {
+    if (platform.is.WEBKIT && platform.is.IOS && platform.majorVersion < 10) {
+      // iOS only does not consider anything from another browsing context keyboard focusable
+      return false;
+    }
+
+    // iframe[tabindex="-1"] and object[tabindex="-1"] inherit the
+    // tabbable demotion onto elements of their browsing contexts
+    if (tabindexValue(frameElement) < 0) {
+      return false;
+    }
+
+    // Webkit and Blink don't consider anything in <object> tabbable
+    const frameNodeName = frameElement.nodeName.toLowerCase();
+    if (frameNodeName === 'object' && (platform.is.BLINK || platform.is.WEBKIT)) {
+      return false;
+    }
   }
 
   const nodeName = element.nodeName.toLowerCase();